Remove unused prefab and resize textures

+++ Assets/Game/Scripts/Enemies/Enemy.cs
@@ -13,7 +13,7 @@ public class Enemy : MonoBehaviour {
 		tag = "Enemy";
 	}
 
-    public virtual void TakeDamage(int _hit)
+    public virtual void TakeDamage(float _hit)
     {
         health -= _hit;
         if (health <= 0)
